chore: move strip-ansi-codes to utils
