+++ packages/cli/src/utils/strip-ansi-codes.ts
